Add workaround for failing pip on Ubuntu/Debian
On a flaky network connections pip tries to restart a process and on
debian/ubuntu then fails with:

    TypeError: unsupported operand type(s) for -=: 'Retry' and 'int'

The original error is explained in
https://stackoverflow.com/a/46970344/1380673 and has something to do with un-
vendored requests/urllib3 on Debian and Ubuntu. Here we simply force pip to
use the newly installed requests/urllib3 instead of the one which came from
the host system.

Closes: #19