Always upgrade packages used in makefile to latest version
The problem was that we installed these packages outside of the requirement.txt file but no
code actually triggered an update.

The pip upgrade was actually part of the venv creation but got deleted at one point
when the pip pinning workaround was added and then removed.

Also switch the pip call to use the more recent `python -m pip` way which in this case is
probably unnecessary as we call a specific `.venv/bin/pip` and not via `pip`
(which can be anything in $PATH).